chore: fix clippy issues

+++ src/mqtt.rs
@@ -225,7 +225,7 @@ pub fn run_mqtt_daemon(
             let msg = paho_mqtt::Message::new(
                 format!("{}/JSON", config.topic),
                 payload,
-                config.qos as i32,
+                config.qos,
             );
             cli.publish(msg)
                 .with_context(|| "Cannot publish MQTT message")?;
